Testing the add timer cell action
The test works, but use of scan to handle the table data means that it needs a viewDidLoad before the taps will do anything, which I’m not super keen on. Scan for this may be a poor choice and short lived anyway so I’ll not get hung up on this right now.